# grpc/article.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "article.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/article": {
      "get": {
        "operationId": "Find",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcArticles"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "subtitle",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "content",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "statusId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "profileId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "ArticleService"
        ]
      },
      "post": {
        "operationId": "Create",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcArticle"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/grpcArticle"
            }
          }
        ],
        "tags": [
          "ArticleService"
        ]
      }
    },
    "/article/{id}": {
      "get": {
        "operationId": "FindById",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcArticle"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "ArticleService"
        ]
      },
      "delete": {
        "operationId": "DeleteById",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcArticle"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "ArticleService"
        ]
      },
      "patch": {
        "operationId": "Update",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/grpcArticle"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/grpcArticleId"
            }
          }
        ],
        "tags": [
          "ArticleService"
        ]
      }
    }
  },
  "definitions": {
    "grpcArticle": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "title": {
          "type": "string"
        },
        "subtitle": {
          "type": "string"
        },
        "content": {
          "type": "string"
        },
        "statusId": {
          "type": "integer",
          "format": "int32"
        },
        "profileId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "grpcArticleId": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "grpcArticles": {
      "type": "object",
      "properties": {
        "offset": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/grpcArticle"
          }
        }
      }
    }
  }
}